requirements.txt
opencv-python
face-recognition
numpy
pyserial
flask